dashboard page
This is the dashboard page. It will be the first page that the user sees when they log in. It will show the user's journals and allow them to create new journals.
As well as showing user's journals, It also will have a search bar that will allow the user to search for journals and show user their current reading streak to track their reading habits.